get shres code building with cc65 #70

+++ src/shres.c
@@ -1,9 +1,9 @@
-#include "shres.h"
+#include "mega65/shres.h"
 
 /* Open SYSPART Shared Resource area.
    Return 0 if success, or non-zero if error.
 */
-char shopen(char *resource_name,unsigned long long required_flags, struct shared_resource *file_handle)
+char shopen(char *resource_name,unsigned long required_flags, struct shared_resource *file_handle)
 {
   return 1;
 }
@@ -16,7 +16,7 @@ char shread(unsigned char *ptr, unsigned int count, struct shared_resource *file
   return 1;
 }
 
-char shseek(struct shared_resource *,unsigned long long offset, unsigned char whence)
+char shseek(struct shared_resource *,unsigned long offset, unsigned char whence)
 {
   return 1;
 }
@@ -26,7 +26,7 @@ shared_resource_dir shdopen()
   return 0xffff;
 }
 
-char sddread(unsigned long long required_flags, shared_resource_dir *directory_handle, struct shared_resource *dirent)
+char sddread(unsigned long required_flags, shared_resource_dir *directory_handle, struct shared_resource *dirent)
 {
   return 1;
 }
